backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
sqlalchemy==2.0.25
pysqlite3==0.5.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
python-multipart==0.0.6
pdfkit==1.0.0
python-dotenv==1.0.0
openai==1.3.7
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
werkzeug==3.0.1
requests==2.31.0
Babel==2.14.0

# Calculation Pipeline Dependencies
networkx==3.2.1
pint==0.23
scipy==1.11.4
sympy==1.12
